+++ packages/core/src/render3/index.ts
@@ -6,7 +6,7 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {createComponentRef, getHostElement, getRenderedText, markDirty, renderComponent, whenRendered} from './component';
+import {createComponentRef, getHostElement, getRenderedText, renderComponent, whenRendered} from './component';
 import {NgOnChangesFeature, PublicFeature, defineComponent, defineDirective, definePipe} from './definition';
 import {InjectFlags} from './di';
 import {ComponentDef, ComponentTemplate, ComponentType, DirectiveDef, DirectiveDefFlags, DirectiveType} from './interfaces/definition';
@@ -65,6 +65,7 @@ export {
   embeddedViewStart as V,
   embeddedViewEnd as v,
   detectChanges,
+  markDirty,
 } from './instructions';
 
 export {
@@ -113,7 +114,6 @@ export {
   createComponentRef,
   getHostElement,
   getRenderedText,
-  markDirty,
   renderComponent,
   whenRendered,
 };
